refatoração codigo/ implementação de regras de negocio

+++ case-entregas/src/main/java/com/example/caseDTI/entregas/case_entregas/entity/Drone.java
@@ -19,7 +19,7 @@ public class Drone {
     private Long id;
 
     @Column
-    private Integer pesoSuportado;
+    private Double pesoSuportado;
 
     @Column
     private Integer autonomiaVoo;
